% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_power.R
\name{compute_power}
\alias{compute_power}
\alias{compute_power.simu_ph_base}
\title{Compute the statistical power of a treatment effect}
\usage{
compute_power(x, ...)

\method{compute_power}{simu_ph_base}(
  x,
  n_reps,
  n_samples,
  formulae,
  get_cindex = TRUE,
  get_lambda = TRUE,
  seed = NULL,
  n_cores = 1,
  ...
)
}
\arguments{
\item{x}{An object of a subclass of \link{simu_ph_base}.}

\item{...}{further arguments passed to other methods.}

\item{n_reps}{Number of repetitions of the trial simulation to estimate the
statistical power of the treatment effect.}

\item{n_samples}{Number of samples in each trial simulation.}

\item{formulae}{A (named) list of formulae specifying the group of covariates
used in the estimation of treatment effect. Each formula will be passed to
\code{\link[survival:coxph]{survival::coxph()}}, and should be of the form \code{Surv(time, event) ~ covariate_1 + covariate_2 + ...}.}

\item{get_cindex}{Logical. If TRUE, estimate the concordance for each of the
formula in \code{formulae} in the control arm (treatment == 0) of the trial.}

\item{get_lambda}{Logical. If TRUE, estimate the cumulative incidence of the
event at the end of the trial in the control arm.}

\item{seed}{A random seed for reproducibility.}

\item{n_cores}{Number of cores to be used in parallel computing.}
}
\value{
A data frame
}
\description{
Compute the statistical power of a treatment effect from a simulated dataset
for a set of formulae, compute optionally also the concordance and cumulative
incidence.
}
\examples{
simu <- simu_ph_weibull(
  shape = 0.5, hr_trt = 0.4, dropout = 0.1, t_cutoff = 5,
  beta = 1, kappa = -1, inclusion = 0.8
)
formulae <- list(
  base = survival::Surv(time, event) ~ 1,
  adj = survival::Surv(time, event) ~ X
)
res <- compute_power(simu, n_reps = 10, n_samples = 500, formulae = formulae)
}
