% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rsquared.R
\name{compute_rsquared}
\alias{compute_rsquared}
\title{Compute various measures of R-squared of the proportional hazard model for a
given dataset and group of covariates}
\usage{
compute_rsquared(data, f, cs_only = FALSE)
}
\arguments{
\item{data}{A data frame of survival data}

\item{f}{A formula describing the group of covaraites}

\item{cs_only}{Logical. If TRUE, only the Cox-Snell R-squared is computed.}
}
\value{
A data frame containing all measures of R-squared
}
\description{
Compute various measures of R-squared of the proportional hazard model for a
given dataset and group of covariates
}
\references{
\itemize{
\item \code{r2_d}: Royston P, Sauerbrei W. A new measure of prognostic separation in
survival data. Stat Med. 2004 Mar 15;23(5):723-48. doi: 10.1002/sim.1621.
PMID: 14981672.
\item \code{r2_cs}, \code{rho_k}, \code{r2_pm}, \code{r2_r}, \code{rho_wa}: Royston, P. (2006). Explained
Variation for Survival Models. The Stata Journal, 6(1), 83–96.
https://doi.org/10.1177/1536867X0600600105
\item \code{r2_xoq}: Ronghui Xu & John O' quigley (1999) A. R.2 type measure of
dependence for proportional hazards models, Journal of Nonparametric
Statistics, 12:1, 83-107, DOI: 10.1080/10485259908832799
}
}
