% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovadjustSim-package.R
\docType{package}
\name{CovadjustSim-package}
\alias{CovadjustSim}
\alias{CovadjustSim-package}
\title{CovadjustSim: Simulation Study of Covariate Adjustment in Clinical Trials}
\description{
A study of the effect of covariate adjustment in clinical trials, based on parametric and semi-synthetic simulations. This package is intended to reproduce the results presented in the article "More efficient and inclusive time-to-event trials with covariate adjustment: a simulation study".
}
\author{
\strong{Maintainer}: Honghao Li \email{honghao.li@owkin.com}

Authors:
\itemize{
  \item Raphaëlle Momal \email{raphaelle.momal@owkin.com}
  \item Félix Balazard \email{felix.balazard@owkin.com}
}

}
\keyword{internal}
