% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eligibility_parametric}
\alias{eligibility_parametric}
\title{Results of parametric study of the effect of covariate adjustment and
eligibility criteria on the statisitical power of the treatment effect
estimation}
\format{
\subsection{\code{eligibility_parametric}}{

A data frame with 8 columns:
\describe{
\item{n_samples}{Number of samples in the whole trial population}
\item{n_events}{Nubmer of events in the whole trial population}
\item{lambda}{Cumulative incidence at 5 years in the control arm}
\item{cindex_base}{Concordance in the control arm when not adjusted on any
covariate}
\item{cindex_adj}{Concordance in the control arm when adjusted on the
standard normal covariate}
\item{power_base}{Statistical power of the treatment effect estimation when
not adjusted on any covariate}
\item{power_adj}{Statistical power of the treatment effect estimation when
adjusted on the standard normal covariate}
\item{inclusion}{Quantile of the standard normal covariate. Samples with
covariate within the quantile are included in the trial.}
}
}
}
\usage{
eligibility_parametric
}
\description{
The cohorts are generated by a fully parametric model based on Weibull
distribution and proportional hazard model, with a standard normal covariate.
Two eligibility criteria based on the quantile of the covariate are studied:
100\% and 50\%. Two sets of adjustments are studied: no adjustment and
adjustment on the standard normal covariate. Together there are four
combinations. For each combination, the dependency of the statistical power
of the treatment effect estimation on the number of events in the whole trial
population is obtained around the 80\% power common standard.
}
\seealso{
\link{simu_ph_weibull}
}
\keyword{datasets}
