% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eligibility_semisynthetic}
\alias{eligibility_semisynthetic}
\title{Results of semi-synthetic study of the effect of covariate adjustment and
eligibility criteria on the statisitical power of the treatment effect
estimation}
\format{
\subsection{\code{eligibility_semisynthetic}}{

A data frame with 8 columns:
\describe{
\item{n_samples}{Number of samples in the whole trial population}
\item{n_events}{Nubmer of events in the whole trial population}
\item{lambda}{Cumulative incidence at 5 years in the control arm}
\item{cindex_clinic}{Concordance in the control arm when adjusted on the
clinical group}
\item{cindex_hcc}{Concordance in the control arm when adjusted on the
clinical+HCC group}
\item{power_clinic}{Statistical power of the treatment effect estimation
when adjusted on the clinical group}
\item{power_hcc}{Statistical power of the treatment effect estimation when
adjusted on the clinical+HCC group}
\item{inclusion}{Percentage of patients in the TCGA dataset that are
eligible to be included in the trial}
}
}
}
\usage{
eligibility_semisynthetic
}
\description{
The cohorts are generated semi-synthetically based on a TCGA dataset.
Three eligibility criteria with minimal, intermediate and maximal
restrictions are studied, corresponding to 100\%, 84\% and 55\% of the samples
in the TCGA dataset. Two sets of adjustment are studied: clinical (\code{stage},
\code{ECOG_SCORE}) and clinical + HCC (\code{stage}, \code{ECOG_SCORE}, \code{predOS}), where
\code{predOS} is a prognostic variable coming from a deep model trained with an
external dataset. Together there are six combinations. For each combination,
the dependency of the statistical power of the treatment effect estimation on
the number of events in the whole trial population is obtained around the 80\%
power common standard.
}
\keyword{datasets}
