% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{estimate_sample_size_reduction}
\alias{estimate_sample_size_reduction}
\title{Estimate the reduction in sample size required to reach 80\% statistical power
when adjusting for a covariate in a clinical trial.}
\usage{
estimate_sample_size_reduction(
  simu,
  formula1,
  begin1,
  end1,
  formula2,
  begin2,
  end2,
  n_reps = 1000,
  n_points = 6,
  seed = NULL,
  n_cores = 1
)
}
\arguments{
\item{simu}{An object of class \link{simu_ph_weibull}}

\item{formula1}{The baseline formula to be analysed in \code{simu}}

\item{begin1, end1, begin2, end2}{Values of \code{n_samples} specifying the range of
the list of \code{n_samples} for each formula, often fixed by \code{\link[=power_skim]{power_skim()}}.}

\item{formula2}{The formula to be compared with \code{formula1} to estimate the
reduction in sample size to reach 80\% power}

\item{n_reps}{Number of repeated experiments to estimate the statistical
power of the treatment effect.}

\item{n_points}{Minimum number of sample size points used to interpolate the
sample size that gives statistical power of 80\%.}

\item{seed}{A random seed for reproducibility.}

\item{n_cores}{Number of cores to be used in parallel computing.}
}
\value{
A data frame including the average cumulative incidence \code{lambda}, the
average concordance \code{cindex} for each formula, the number of samples
corresponding to 80\% statistical power for each formula, and the estimated
sample size reduction.
}
\description{
For each formula, a list of \code{n_samples} will be generated. For each
\code{n_samples}, statistical power of the treatment effect is estimated. Then for
each formula, the \code{power ~ n_samples} curve will be interpolated to estimate
the sample size required to achieve 80\% power.
}
\examples{
simu <- simu_ph_weibull(
  shape = 0.5, hr_trt = 0.4, dropout = 0.01, t_cutoff = 5, beta = 0.58,
  kappa = -1.53
)
formula1 <- survival::Surv(time, event) ~ 1
formula2 <- survival::Surv(time, event) ~ X

estimate_sample_size_reduction(simu,
  formula1, 80, 200,
  formula2, 70, 170,
  n_reps = 30, n_points = 5, seed = 1
)
}
\seealso{
\code{\link[=power_skim]{power_skim()}}
}
