% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fix_dist_params}
\alias{fix_dist_params}
\title{Fix \code{kappa} and \code{beta} parameters for multiple experiment settings}
\usage{
fix_dist_params(param_grid, ..., seed = NULL, n_cores = 1)
}
\arguments{
\item{param_grid}{A data frame. Must contains the following columns:
\code{shape}, \code{dropout}, \code{t_cutoff}, \code{cindex}, \code{lambda}.}

\item{...}{Additional arguments passed on to \code{get_mapping_grid}.}

\item{seed}{A random seed for reproducibility.}

\item{n_cores}{Number of cores to be used in parallel computing.}
}
\value{
A data frame of \code{param_grid} with two additional columns: \code{beta} and
\code{kappa}.
}
\description{
The relations between (\code{cindex}, \code{lambda}) and (\code{kappa}, \code{beta}) are
determined given a tuple (\code{shape}, \code{dropout}, \code{t_cutoff}).
For each of these tuples, the corresponding (\code{kappa}, \code{beta}) pairs for
multiple (\code{cindex}, \code{lambda}) pairs can be interpolated with the same model.
First, a fine grid of \code{beta} and \code{kappa} (ranges depend on \code{shape} and the
covariate, and are fixed manually) is generated, the values of each (cindex,
lambda) pair is estimated for each point on the grid. Then the value of
\code{beta} corresponding to the desired \code{cindex} is interpolated (\code{cindex} does
not depend on \code{kappa}) with a polynomial. Then given the values of \code{beta} and
\code{lambda}, the value of \code{kappa} is interpolated with \code{\link[interp:interp]{interp::interp()}}.
}
\examples{
param_grid <- tidyr::expand_grid(
  shape = 0.5,
  dropout = 0.01,
  t_cutoff = 5,
  n_reps = 10000,
  hr_trt = 0.7,
  cindex = c(0.75, 0.85),
  lambda = seq(0.1, 0.6, 0.1)
)
beta_list <- c(0.001, seq(0.1, 3, length.out = 9))
kappa_list <- seq(-8 + 0.4, 4, length.out = 7)
fix_dist_params(
  param_grid, beta_list = beta_list, kappa_list = kappa_list,
  n_samples = 500, seed = 42
)
}
