% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_skim.R
\name{interpolate_power}
\alias{interpolate_power}
\title{Find values of \code{n_samples} for given values of statistical power}
\usage{
interpolate_power(df_power, power_targets)
}
\arguments{
\item{df_power}{A data frame of \code{power}, \code{n_samples} and \code{n_events}}

\item{power_targets}{Values of \code{power} for which corresponding \code{n_samples}
and \code{n_events} will interpolated}
}
\description{
Often combined with the function \code{\link[=power_skim]{power_skim()}} to get a quick estimation of
the required sample size and number of events for a treatment effect
estimation to reach a predefined power.
}
\seealso{
\code{\link[=power_skim]{power_skim()}}
}
