% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parametric_study}
\alias{parametric_study}
\title{Results of parametric study of the reduction of required sample size to reach
80\% statistical power of the treatment effect estimation when using covariate
adjustment.}
\format{
An object of class \code{data.frame} with 432 rows and 26 columns.
}
\usage{
parametric_study
}
\description{
The dependence of the reduction on various parameters are studied, for each
combination of parameters, the reduction is also compared with various
measures of r-squared for time-to-event data.
}
\details{
A data frame with 432 rows and 26 columns. Some variables of interest are:
\describe{
\item{shape,dropout,t_cutoff,hr_trt,cindex0,lambda0}{Parameters defining a
simulation, see \link{simu_ph_weibull} and \code{\link[=fix_dist_params]{fix_dist_params()}} for details.}
\item{r2_obs_80}{Estimated reduction of required sample size to reach 80\%
statistical power}
\item{r2_\emph{,rho_}}{Various r-squared measures for time-to-event data}
...
}
}
\keyword{datasets}
