% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_skim.R
\name{power_skim}
\alias{power_skim}
\alias{power_skim.simu_ph_base}
\title{Skim through the power range of a simulation}
\usage{
power_skim(x, ...)

\method{power_skim}{simu_ph_base}(
  x,
  f,
  n_reps = 300,
  n_points = 7,
  n_init = 20,
  power_stop = 0.9,
  ...
)
}
\arguments{
\item{x}{An object of a subclass of \link{simu_ph_base}}

\item{...}{further arguments passed to other methods.}

\item{f}{A formula specifying the covariates}

\item{n_reps}{Number of repetitions of the trial simulation at each
\code{n_samples} to estimate the statistical power of the treatment effect.
Intended to be small to keep the speed of computation.}

\item{n_points}{Minimal number of \code{n_samples} points to be examined}

\item{n_init}{starting point of \code{n_samples}}

\item{power_stop}{Upper limit of power to stop the skim}
}
\value{
A data frame of \code{n_samples}, \code{n_events} and \code{power}
}
\description{
Loosely compute the statistical power of the treatment effect at a series of
\code{n_samples} points, until the upper limit of power is reached. At each step,
the increment of \code{n_samples} is estimated adaptively.
}
\seealso{
\code{\link[=interpolate_power]{interpolate_power()}}
}
