% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simu_ph_base}
\alias{simu_ph_base}
\title{Generic S3 class of basic trial simulation based on proportional hazard model}
\usage{
simu_ph_base(class = character())
}
\arguments{
\item{class}{A string specifying the subclass}
}
\value{
An object of class \code{simu_ph_base} or the derived subclass
}
\description{
The class has no member and a method \code{\link[=compute_power]{compute_power()}}. Derived classes
should implement the method \code{\link[=generate_trial_data]{generate_trial_data()}}, and optionally the
method \code{\link[=compute_power]{compute_power()}}.
}
\seealso{
\link{simu_ph_weibull}
}
\keyword{internal}
